extract about scroll to comp
