+++ src/components/AboutScroll/styles.scss
@@ -1,28 +1,26 @@
-aside {
-  font-size: 4vw;
-  /* 	text-align: right; */
-}
-
 .container {
   display: flex;
   /* 	gap: 30px; */
+
+  .left {
+    width: 50%;
+  }
+
+  .right {
+    display: flex;
+    padding: 4vw 0;
+
+    .ruler {
+      width: 5px;
+      background-color: black;
+      height: 100vh;
+      margin-right: 1.5vw;
+      height: auto;
+    }
+  }
 }
 
-.ruler {
-  width: 5px;
-  background-color: black;
-  /* 	height: 100vh; */
-}
-
-main {
-  width: 100vw;
-}
-
-.left {
-  width: 50%;
-}
-
-.right {
-  /* 	display: flex; */
-  padding: 4vw 0;
+aside {
+  font-size: 4vw;
+  /* 	text-align: right; */
 }
