+++ src/components/Navbar/Navbar.scss
@@ -11,6 +11,7 @@
   left: 0;
   right: 0;
   bottom: 0;
+  z-index: 100;
 }
 
 .right-items {
